chore: cleanup and added comments
