--- src/lib/Option.svelte
@@ -1,8 +1,0 @@
-<script>
-    export let label;        // For the <summary>
-</script>
-
-  <label class="label">
-    <input type="checkbox" checked="checked" class="checkbox" />
-    {label}
-  </label>
