+++ src/routes/digital-catalog/+page.js
@@ -3,6 +3,7 @@ import { getBooks, getFilterOptions, getPreviewFilters } from '$lib';
 
 export async function load({ url, fetch }) {
 
+	// #region Pagination
 	let resultsPage = parseInt(url.searchParams.get('results-page')) || 1;
 	const resultsPageAction = url.searchParams.get('results-page-action');
 	if (resultsPageAction === 'next') {
@@ -10,26 +11,6 @@ export async function load({ url, fetch }) {
 	} else if (resultsPageAction === 'previous') {
 		resultsPage--
 	}
-
-	let startYear = parseInt(url.searchParams.get('start-year'));
-	let endYear = parseInt(url.searchParams.get('end-year'));
-	if (!startYear && !endYear) {
-		// If start and endYear are unset, try to use century instead
-		const century = parseInt(url.searchParams.get('century'));
-		startYear = (century - 1) * 100;
-		endYear = century * 100;
-	}
-
-	const filter = {
-		searchTerm: url.searchParams.get('q'),
-		author: url.searchParams.get('author'),
-		place: url.searchParams.get('place'),
-		digitalized: url.searchParams.get('digitalized') === 'true',
-		startYear,
-		endYear
-	}
-
-	const { books, totalResults } = await getBooks(resultsPage, filter, fetch);
 
 	let authorsPage = parseInt(url.searchParams.get('authors-page')) || 1;
 	const authorsPageAction = url.searchParams.get('authors-page-action');
@@ -46,6 +27,29 @@ export async function load({ url, fetch }) {
 	} else if (placesPageAction === 'previous') {
 		placesPage--;
 	}
+	// #endregion Pagination
+
+	// #region Calculate Year
+	let startYear = parseInt(url.searchParams.get('start-year'));
+	let endYear = parseInt(url.searchParams.get('end-year'));
+	if (!startYear && !endYear) {
+		// If start and endYear are unset, try to use century instead
+		const century = parseInt(url.searchParams.get('century'));
+		startYear = (century - 1) * 100;
+		endYear = century * 100;
+	}
+	// #endregion Calculate Year
+
+	const filter = {
+		searchTerm: url.searchParams.get('q'),
+		author: url.searchParams.get('author'),
+		place: url.searchParams.get('place'),
+		digitalized: url.searchParams.get('digitalized') === 'true',
+		startYear,
+		endYear
+	}
+
+	const { books, totalResults } = await getBooks(resultsPage, filter, fetch);
 
 	const {options: authors, total: totalAuthors} = await getFilterOptions('auteur', 66, authorsPage, fetch);
 	const {options: places, total: totalPlaces} = await getFilterOptions('plaats_van_uitgave', 66, placesPage, fetch);
